refactor: remove action models
