+++ src/main/java/local/uniclog/model/actions/impl/ActionWhile.java
@@ -1,7 +1,7 @@
-package local.uniclog.model.actions;
+package local.uniclog.model.actions.impl;
 
-import local.uniclog.model.ActionType;
-import local.uniclog.model.ActionsInterface;
+import local.uniclog.model.actions.ActionType;
+import local.uniclog.model.actions.ActionsInterface;
 import local.uniclog.utils.DataUtils;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,7 +11,6 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.TimeUnit;
 
 import static java.lang.String.format;
 
@@ -20,14 +19,13 @@ import static java.lang.String.format;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Sleep implements ActionsInterface {
+public class ActionWhile implements ActionsInterface {
     @Builder.Default
-    private Long time = 0L;
+    private Integer count = 0;
 
     @Override
     public void execute(String... args) throws InterruptedException {
         log.debug("{}", this);
-        TimeUnit.MILLISECONDS.sleep(time);
     }
 
     @Override
@@ -38,8 +36,8 @@ public class Sleep implements ActionsInterface {
     }
 
     private void setFieldValue(String key, String value) {
-        if ("time".equals(key)) {
-            setTime(DataUtils.getLong(value, 0));
+        if ("count".equals(key)) {
+            setCount(DataUtils.getInteger(value, 0));
         } else {
             log.debug("Field: {}, not set: {}", key, value);
         }
@@ -47,11 +45,11 @@ public class Sleep implements ActionsInterface {
 
     @Override
     public String toString() {
-        return format("%s [time=%s]", getType().name(), time);
+        return format("%s [count=%d]", getType().name(), count);
     }
 
     @Override
     public ActionType getType() {
-        return ActionType.SLEEP;
+        return ActionType.WHILE;
     }
 }
