+++ src/main/java/local/uniclog/model/actions/impl/Default.java
@@ -1,6 +1,6 @@
-package local.uniclog.model.actions;
+package local.uniclog.model.actions.impl;
 
-import local.uniclog.model.ActionsInterface;
+import local.uniclog.model.actions.ActionsInterface;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.Map;
